Refactor: summarize repeated functions
